Created BaseMaskTextWatcher with core logic to format as TextWatcher
